+++ api/src/main/java/net/jqwik/api/arbitraries/MapArbitrary.java
@@ -1,6 +1,6 @@
 package net.jqwik.api.arbitraries;
 
-import java.util.stream.*;
+import java.util.*;
 
 import org.apiguardian.api.*;
 
@@ -8,36 +8,36 @@ import static org.apiguardian.api.API.Status.*;
 
 /**
  * Fluent interface to add functionality to arbitraries that generate instances
- * of type {@linkplain Stream}
+ * of type {@linkplain Map}
  */
 @API(status = MAINTAINED, since = "1.3.2")
-public interface StreamArbitrary<T> extends StreamableArbitrary<T, Stream<T>> {
+public interface MapArbitrary<K, V> extends SizableArbitrary<Map<K, V>> {
 
 	/**
 	 * Fix the size to {@code size}.
 	 *
-	 * @param size The size of the generated stream
+	 * @param size The size of the generated map
 	 * @return new arbitrary instance
 	 */
 	@Override
-	default StreamArbitrary<T> ofSize(int size) {
+	default MapArbitrary<K, V> ofSize(int size) {
 		return ofMinSize(size).ofMaxSize(size);
 	}
 
 	/**
 	 * Set lower size boundary {@code minSize} (included).
 	 *
-	 * @param minSize The minimum size of the generated stream
+	 * @param minSize The minimum size of the generated map
 	 * @return new arbitrary instance
 	 */
-	StreamArbitrary<T> ofMinSize(int minSize);
+	MapArbitrary<K, V> ofMinSize(int minSize);
 
 	/**
 	 * Set upper size boundary {@code maxSize} (included).
 	 *
-	 * @param maxSize The maximum size of the generated stream
+	 * @param maxSize The maximum size of the generated map
 	 * @return new arbitrary instance
 	 */
-	StreamArbitrary<T> ofMaxSize(int maxSize);
+	MapArbitrary<K, V> ofMaxSize(int maxSize);
 
 }